Organize project structure by relocating EscapeRoomListPage component
